growth_agent/requirements.txt
telethon>=1.36
asyncpg>=0.30
pydantic>=2.0
httpx>=0.27
loguru>=0.7
python-dotenv>=1.0
